// nextJs/src/types/movie.ts
export type Movie = {
    id: string
    posterLink: string | null
    seriesTitle: string
    releasedYear: string | null
    certificate: string | null
    runtime: string | null
    genre: string | null
    imdbRating: number | null
    overview: string | null
    metaScore: number | null
    director: string | null
    star1: string | null
    star2: string | null
    star3: string | null
    star4: string | null
    noOfVotes: number | null
    gross: string | null
}